requirements.txt
aws-lambda-powertools[aws-sdk]
langchain==0.0.271
openai==0.27.9
loguru==0.7.0
pydantic==1.10.12
tiktoken==0.4.0
unstructured[epub]==0.10.5
pypdf==3.11.1
python-magic==0.4.27
chromadb==0.4.7
pysqlite3-binary==0.5.1